Remove logic for progressive web app
The service worker caches all the static assets and loads them from
memory, then updates them in the background. But this comes at the cost
of updates not going immediately. This causes problems when contracts
between backend and frontend is not the same. Also, for some reason
/ -> /locks -> github -> / keeps looping due to the service worker, I do
not understand why this happens.

So, for now I am disabling this feature, maybe I can revisit this later.